[Chore][Manifest] address review feedback
Move _ArgreduceWorkload from bench_argreduce.py into workloads/ops/argreduce.py
as ArgmaxTest/ArgminTest, following the shared workload layer pattern.
